# ajnaapi/openapi.yaml
openapi: 3.0.1
info:
  title: API-AJNA
  description: API para acesso de outras aplicações a dados do AJNA
  version: "1.0b"
  termsOfService: "https://ajna.labin08.rf08.srf/"
  contact:
    url: "https://ajna.labin08.rf08.srf/"
    name: "Receita Federal"
    email: "ivan.brasilico@rfb.gov.br"
  license:
    name: Receita Federal C-level site
    url: "https://ajna.labin08.rf08.srf/site"
...
servers:
  - url: /ajnaapi/api
...
paths:
...
  /acessoveiculo/{id}:
    get:
      summary: Retorna objeto
      tags:
        - API Recintos
      parameters:
        - name: id
          in: path
          description: Id do Evento
          required: true
          schema:
            type: string
      responses:
        200:
          description: Evento Acesso Veiculo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcessoVeiculo'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /acessoveiculo:
    post:
      summary: Recebe Evento acessoveiculo
      tags:
        - API Recintos
      operationId: authentication.generate_token
      requestBody:
        description: Conteúdo do Evento
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcessoVeiculo'
      responses:
        201:
          description: Resposta em caso de sucesso
          content:
            application/json:
              schema: {}
        500:
          description: Erro
          content:
            application/json:
              schema: {}
      security:
        - jwt: ['secret']
  /pesagemveiculo/{id}:
    get:
      tags:
        - API Recintos
      parameters:
        - name: id
          in: path
          description: Id do Evento
          required: true
          schema:
            type: string
      responses:
        200:
          description: Evento Pesagem Veiculo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PesagemVeiculo'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /pesagemveiculo:
    post:
      summary: Recebe Evento pesagemveiculo
      tags:
        - API Recintos
      operationId: authentication.generate_token
      requestBody:
        description: Conteúdo do Evento
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PesagemVeiculo'
      responses:
        201:
          description: Resposta em caso de sucesso
          content:
            application/json:
              schema: {}
        500:
          description: Erro
          content:
            application/json:
              schema: {}
      security:
        - jwt: ['secret']
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: authentication.decode_token
  schemas:
...
    AcessoVeiculo:
      type: object
      properties:
        id:
          type: number
        cnpjTransportador:
          type: string
        codRecintoDestino:
          type: integer
        contingencia:
          type: boolean
        cpfOperOcor:
          type: string
        cpfOperReg:
          type: string
        direcao:
          type: string
        dtHrAgenda:
          type: string
        dtHrOcorrencia:
          type: string
        dtHrRegistro:
          type: string
        gate:
          type: string
        idEvento:
          type: string
        listaConteineresUld:
          properties: listaConteineresUld
          type: object
          $ref: '#components/schemas/listaConteinerUld'
        login:
          type: string
        mercadoria:
          type: string
        modal:
          type: string
        motorista_cpf:
          type: string
        motorista_nome:
          type: string
        motorista_protocoloCredenciamento:
          type: string
        nmtransportador:
          type: string
        ocrPlaca:
          type: boolean
        oogDimensao:
          type: boolean
        oogPeso:
          type: boolean
        operacao:
          type: string
        placa:
          type: string
        protocoloAgenda:
          type: string
        protocoloEventoRetifCanc:
          type: string
        recinto:
          type: string
        tipoGranel:
          type: string
...
    PesagemVeiculo:
      type: object
      properties:
        id:
          type: number
        balanca:
          type: string
        capturaAutoPeso:
          type: boolean
        contingencia:
          type: boolean
        cpfOperOcor:
          type: string
        cpfOperReg:
          type: string
        dtHrOcorrencia:
          type: string
        dtHrRegistro:
          type: string
        dutos:
          type: string
        idEvento:
          type: string
        listaSemirreboque:
          properties: listaSemirreboque
          type: object
          $ref: '#components/schemas/Semirreboque'
        numConteinerUld:
          type: string
        pesoBrutoBalanca:
          type: number
        pesoBrutoManifesto:
          type: number
        placa:
          type: string
        protocoloEventoRetifCanc:
          type: string
        recinto:
          type: string
        taraConjunto:
          type: number
        taraConteinerUld:
          type: number
        tipoConteinerUld:
          type: string
        volume:
          type: number
...